+++ src/virtual_keyboard.rs
@@ -58,7 +58,7 @@ pub struct VirtualKeyboard {
 
     base_layer: NodeIndex<u8>,
     locked_layer: NodeIndex<u8>,
-    active_layer: NodeIndex<u8>,
+    layer_history: Vec<NodeIndex<u8>>,
 
     pressed_keys: HashMap<u16, Option<KeyAction>>,
     pressed_modifiers: Vec<u16>,
@@ -105,7 +105,7 @@ impl VirtualKeyboard {
             modifiers_scan_codes,
             base_layer,
             locked_layer: base_layer,
-            active_layer: base_layer,
+            layer_history: vec![base_layer],
             pressed_keys: HashMap::new(),
             pressed_modifiers: Vec::new(),
         })
@@ -140,7 +140,7 @@ impl VirtualKeyboard {
         }
 
         let mut layer_activations = self.layer_activations();
-        self.active_layer = if let Some(active_layer) = layer_activations.next() {
+        let active_layer = if let Some(active_layer) = layer_activations.next() {
             // Lock the layer if we find a second sequence for this layer
             // Example: Both shift key pressed to lock the shift layer (caps lock functionality).
             if layer_activations.any(|layer| layer == active_layer) {
@@ -150,12 +150,42 @@ impl VirtualKeyboard {
                 // Update graph with the locked layer as new base layer.
                 reverse_edges(&mut self.layer_graph, self.base_layer, active_layer);
 
+                // Jump back in history if this layer was locked before.
+                if let Some(idx) = self
+                    .layer_history
+                    .iter()
+                    .position(|layer| *layer == active_layer)
+                {
+                    self.layer_history.drain(idx + 1..);
+                }
+
                 self.locked_layer = active_layer;
             }
 
             active_layer
         } else {
             self.locked_layer
+        };
+
+        // When the active layer has changed search for it in the history but stop at the locked layer.
+        let mut layer_idx = None;
+        for idx in (0..self.layer_history.len()).rev() {
+            if self.layer_history[idx] == active_layer {
+                layer_idx = Some(idx);
+                break;
+            }
+
+            if self.layer_history[idx] == self.locked_layer {
+                break;
+            }
+        }
+
+        // Update layer history.
+        if let Some(idx) = layer_idx {
+            // Remove all layers “newer” than the active layer.
+            self.layer_history.drain(idx + 1..);
+        } else {
+            self.layer_history.push(active_layer);
         }
     }
 
@@ -167,16 +197,14 @@ impl VirtualKeyboard {
         // may not be the same if the layer has changed in between.
         let action = self.pressed_keys.remove(&scan_code).unwrap_or_else(|| {
             let key = self.layout.get_key(scan_code);
-            key.action_on_layer(self.active_layer.index() as _)
 
-            // let mut action = None;
-            // for layer in layer_history {
-            //     if let Some(action) = key.action_on_layer(layer) {
-            //         action = Some(action);
-            //         break;
-            //     }
-            // }
-            // action
+            // Get the key action from the current layer. If the key is not available on
+            // the current layer, check the previous layer. Repeat until a action was
+            // found or we run out of layers.
+            self.layer_history
+                .iter()
+                .rev()
+                .find_map(|layer| key.action_on_layer(layer.index() as _))
         });
 
         self.update_modifiers(scan_code, false);
@@ -272,13 +300,13 @@ mod tests {
         assert_eq!(kb.press_key(0x12), Some(Ignore));
         assert_eq!(kb.press_key(0x20), Some(Character('2')));
         assert_eq!(kb.release_key(0x20), Some(Character('2')));
-        assert_eq!(kb.press_key(0x11), None);
+        assert_eq!(kb.press_key(0x11), Some(Ignore));
         assert_eq!(kb.press_key(0x20), Some(Character('2')));
         assert_eq!(kb.release_key(0x20), Some(Character('2')));
         assert_eq!(kb.release_key(0x12), Some(Ignore));
         assert_eq!(kb.press_key(0x20), Some(Character('1')));
         assert_eq!(kb.release_key(0x20), Some(Character('1')));
-        assert_eq!(kb.release_key(0x11), None);
+        assert_eq!(kb.release_key(0x11), Some(Ignore));
         assert_eq!(kb.press_key(0x20), Some(Character('0')));
         assert_eq!(kb.release_key(0x20), Some(Character('0')));
 
@@ -303,8 +331,8 @@ mod tests {
         let mut kb = VirtualKeyboard::new(layout)?;
 
         assert_eq!(kb.press_key(0xE036), Some(VirtualKey(0xA1)));
-        assert_eq!(kb.press_key(0x002A), None);
-        assert_eq!(kb.release_key(0x002A), None);
+        assert_eq!(kb.press_key(0x002A), Some(VirtualKey(0xA0)));
+        assert_eq!(kb.release_key(0x002A), Some(VirtualKey(0xA0)));
         assert_eq!(kb.release_key(0xE036), Some(VirtualKey(0xA1)));
 
         Ok(())
@@ -352,7 +380,7 @@ mod tests {
         assert_eq!(kb.press_key(0xBB), Some(Character('B')));
         assert_eq!(kb.release_key(0xBB), Some(Character('B')));
 
-        // Release leayer b key
+        // Release layer b key
         assert_eq!(kb.release_key(0x0B), Some(Ignore));
 
         // "B" does not exist on base layer
@@ -370,63 +398,156 @@ mod tests {
             .add_modifier(0xA0, "base", "a", None)
             .add_modifier(0x0B, "base", "b", None)
             .add_modifier(0xB0, "base", "b", None)
+            .add_key(0xFF, "base", Character('X'))
             .add_modifier(0x0B, "a", "c", None)
             .add_modifier(0xB0, "a", "c", None)
-            .add_key(0xAA, "a", Character('A'))
+            .add_key(0xFF, "a", Character('A'))
             .add_modifier(0x0A, "b", "c", None)
             .add_modifier(0xA0, "b", "c", None)
-            .add_key(0xBB, "b", Character('B'))
-            .add_key(0xCC, "c", Character('C'));
+            .add_key(0xFF, "b", Character('B'))
+            .add_key(0xFF, "c", Character('C'));
         let layout = layout.build();
         let mut kb = VirtualKeyboard::new(layout)?;
 
         // Lock layer a
         assert_eq!(kb.press_key(0x0A), Some(Ignore));
-        assert_eq!(kb.press_key(0xA0), None);
+        assert_eq!(kb.press_key(0xA0), Some(Ignore));
         assert_eq!(kb.release_key(0x0A), Some(Ignore));
-        assert_eq!(kb.release_key(0xA0), None);
+        assert_eq!(kb.release_key(0xA0), Some(Ignore));
 
         // Test if locked
-        assert_eq!(kb.press_key(0xAA), Some(Character('A')));
-        assert_eq!(kb.release_key(0xAA), Some(Character('A')));
+        assert_eq!(kb.press_key(0xFF), Some(Character('A')));
+        assert_eq!(kb.release_key(0xFF), Some(Character('A')));
 
         // Temp switch back to layer base
-        assert_eq!(kb.press_key(0x0A), None);
-        assert_eq!(kb.press_key(0xAA), None);
-        assert_eq!(kb.release_key(0xAA), None);
-        assert_eq!(kb.release_key(0x0A), None);
+        assert_eq!(kb.press_key(0x0A), Some(Ignore));
+        assert_eq!(kb.press_key(0xFF), Some(Character('X')));
+        assert_eq!(kb.release_key(0xFF), Some(Character('X')));
+        assert_eq!(kb.release_key(0x0A), Some(Ignore));
 
         // Temp switch to layer c
         assert_eq!(kb.press_key(0x0B), Some(Ignore));
-        assert_eq!(kb.press_key(0xCC), Some(Character('C')));
-        assert_eq!(kb.release_key(0xCC), Some(Character('C')));
+        assert_eq!(kb.press_key(0xFF), Some(Character('C')));
+        assert_eq!(kb.release_key(0xFF), Some(Character('C')));
 
         // Lock layer c
-        assert_eq!(kb.press_key(0xB0), None);
-        assert_eq!(kb.release_key(0xB0), None);
+        assert_eq!(kb.press_key(0xB0), Some(Ignore));
+        assert_eq!(kb.release_key(0xB0), Some(Ignore));
 
         // Temp switched to layer a still
-        assert_eq!(kb.press_key(0xAA), Some(Character('A')));
-        assert_eq!(kb.release_key(0xAA), Some(Character('A')));
+        assert_eq!(kb.press_key(0xFF), Some(Character('A')));
+        assert_eq!(kb.release_key(0xFF), Some(Character('A')));
 
         // Check if locked to layer c
         assert_eq!(kb.release_key(0x0B), Some(Ignore));
-        assert_eq!(kb.press_key(0xCC), Some(Character('C')));
-        assert_eq!(kb.release_key(0xCC), Some(Character('C')));
+        assert_eq!(kb.press_key(0xFF), Some(Character('C')));
+        assert_eq!(kb.release_key(0xFF), Some(Character('C')));
 
         // Lock layer base again
-        assert_eq!(kb.press_key(0xB0), None);
-        assert_eq!(kb.press_key(0xA0), None);
+        assert_eq!(kb.press_key(0xB0), Some(Ignore));
+        assert_eq!(kb.press_key(0xA0), Some(Ignore));
         assert_eq!(kb.press_key(0x0A), Some(Ignore));
         assert_eq!(kb.press_key(0x0B), Some(Ignore));
-        assert_eq!(kb.release_key(0xB0), None);
-        assert_eq!(kb.release_key(0xA0), None);
+        assert_eq!(kb.release_key(0xB0), Some(Ignore));
+        assert_eq!(kb.release_key(0xA0), Some(Ignore));
         assert_eq!(kb.release_key(0x0A), Some(Ignore));
         assert_eq!(kb.release_key(0x0B), Some(Ignore));
 
         // Check if locked to layer base
-        assert_eq!(kb.press_key(0xAA), None);
-        assert_eq!(kb.release_key(0xAA), None);
+        assert_eq!(kb.press_key(0xFF), Some(Character('X')));
+        assert_eq!(kb.release_key(0xFF), Some(Character('X')));
+
+        Ok(())
+    }
+
+    #[test]
+    fn transparency() -> anyhow::Result<()> {
+        let mut layout = LayoutBuilder::new();
+        layout
+            .add_modifier(0x0B, "a", "b", None)
+            .add_modifier(0xB0, "a", "b", None)
+            .add_key(0x01, "a", Character('A'))
+            .add_key(0x02, "a", Character('A'))
+            .add_key(0x03, "a", Character('A'))
+            .add_modifier(0x0C, "b", "c", None)
+            .add_modifier(0xC0, "b", "c", None)
+            .add_key(0x01, "b", Character('B'))
+            .add_key(0x02, "b", Character('B'))
+            .add_key(0x01, "c", Character('C'))
+            .add_key(0x04, "c", Character('C'));
+        let layout = layout.build();
+        let mut kb = VirtualKeyboard::new(layout)?;
+
+        // Layer a
+        assert_eq!(kb.press_key(0x01), Some(Character('A')));
+        assert_eq!(kb.release_key(0x01), Some(Character('A')));
+        assert_eq!(kb.press_key(0x02), Some(Character('A')));
+        assert_eq!(kb.release_key(0x02), Some(Character('A')));
+        assert_eq!(kb.press_key(0x03), Some(Character('A')));
+        assert_eq!(kb.release_key(0x03), Some(Character('A')));
+        assert_eq!(kb.press_key(0x04), None);
+        assert_eq!(kb.release_key(0x04), None);
+
+        assert_eq!(kb.press_key(0x0B), Some(Ignore));
+
+        // Layer b
+        assert_eq!(kb.press_key(0x01), Some(Character('B')));
+        assert_eq!(kb.release_key(0x01), Some(Character('B')));
+        assert_eq!(kb.press_key(0x02), Some(Character('B')));
+        assert_eq!(kb.release_key(0x02), Some(Character('B')));
+        assert_eq!(kb.press_key(0x03), Some(Character('A')));
+        assert_eq!(kb.release_key(0x03), Some(Character('A')));
+        assert_eq!(kb.press_key(0x04), None);
+        assert_eq!(kb.release_key(0x04), None);
+
+        assert_eq!(kb.press_key(0x0C), Some(Ignore));
+
+        // Layer c
+        assert_eq!(kb.press_key(0x01), Some(Character('C')));
+        assert_eq!(kb.release_key(0x01), Some(Character('C')));
+        assert_eq!(kb.press_key(0x02), Some(Character('B')));
+        assert_eq!(kb.release_key(0x02), Some(Character('B')));
+        assert_eq!(kb.press_key(0x03), Some(Character('A')));
+        assert_eq!(kb.release_key(0x03), Some(Character('A')));
+        assert_eq!(kb.press_key(0x04), Some(Character('C')));
+        assert_eq!(kb.release_key(0x04), Some(Character('C')));
+
+        // Lock layer c
+        assert_eq!(kb.press_key(0xB0), Some(Ignore));
+        assert_eq!(kb.press_key(0xC0), Some(Ignore));
+
+        // Release all but on modifier to activate layer b
+        assert_eq!(kb.release_key(0xB0), Some(Ignore));
+        assert_eq!(kb.release_key(0x0B), Some(Ignore));
+        assert_eq!(kb.release_key(0x0C), Some(Ignore));
+
+        dbg!(&kb.locked_layer);
+        dbg!(&kb.layer_history);
+
+        // Layer c
+        assert_eq!(kb.press_key(0x01), Some(Character('B')));
+        assert_eq!(kb.release_key(0x01), Some(Character('B')));
+        assert_eq!(kb.press_key(0x02), Some(Character('B')));
+        assert_eq!(kb.release_key(0x02), Some(Character('B')));
+        assert_eq!(kb.press_key(0x03), Some(Character('A')));
+        assert_eq!(kb.release_key(0x03), Some(Character('A')));
+        // Should be transparent to layer c now
+        assert_eq!(kb.press_key(0x04), Some(Character('C')));
+        assert_eq!(kb.release_key(0x04), Some(Character('C')));
+
+        // Lock layer a again
+        assert_eq!(kb.press_key(0xB0), Some(Ignore));
+        assert_eq!(kb.press_key(0x0C), Some(Ignore));
+        assert_eq!(kb.press_key(0x0B), Some(Ignore));
+
+        assert_eq!(kb.press_key(0x01), Some(Character('A')));
+        assert_eq!(kb.release_key(0x01), Some(Character('A')));
+        assert_eq!(kb.press_key(0x02), Some(Character('A')));
+        assert_eq!(kb.release_key(0x02), Some(Character('A')));
+        assert_eq!(kb.press_key(0x03), Some(Character('A')));
+        assert_eq!(kb.release_key(0x03), Some(Character('A')));
+        assert_eq!(kb.press_key(0x04), None);
+        assert_eq!(kb.release_key(0x04), None);
 
         Ok(())
     }